move state as part of api
