+++ cluster/src/main/java/io/scalecube/cluster/leaderelection/StateMachine.java
@@ -1,5 +1,6 @@
 package io.scalecube.cluster.leaderelection;
 
+import io.scalecube.cluster.leaderelection.api.State;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
