+++ cluster/src/main/java/io/scalecube/cluster/leaderelection/api/ElectionEvent.java
@@ -1,7 +1,5 @@
 package io.scalecube.cluster.leaderelection.api;
 
-import io.scalecube.cluster.leaderelection.State;
-
 public class ElectionEvent {
 
   private static ElectionEvent follower = new ElectionEvent(State.FOLLOWER);
