+++ cluster/src/main/java/io/scalecube/cluster/leaderelection/api/ElectionTopic.java
@@ -1,6 +1,5 @@
 package io.scalecube.cluster.leaderelection.api;
 
-import io.scalecube.cluster.leaderelection.State;
 import reactor.core.publisher.Flux;
 
 public interface ElectionTopic {
