+++ cluster/src/main/java/io/scalecube/cluster/leaderelection/api/State.java
@@ -1,4 +1,4 @@
-package io.scalecube.cluster.leaderelection;
+package io.scalecube.cluster.leaderelection.api;
 
 import io.scalecube.cluster.Cluster;
